Add files via upload
Added a patch to force BGSLightingTemplates to load editor IDs to a storage space we provide; patch is written so that it can be generalized to other forms. It would seem that all forms parse the EDID record and pass its contents to virtual method 0x33 SetEditorID; that method is just a no-op for the vast majority of form types. This patch writes into the VTBL to replace the no-op.