Add files via upload
Overhaul to the DetectionInterceptService. Designing it around a handle exchange isn't viable, because a common use case would unavoidably result in zombie handles: if you try to use the service from an ActiveMagicEffect placed on a leveled actor, the effect will be deleted without running OnEffectFinish if the leveled actor respawns before the effect expires; the handle-exchange design did kill registrations on these actors when they respawned, but it could not delete them (because it's impossible for us to know that the only script holding the handle has been deleted).

The new design relies solely on string tags rather than handles, and registrations are not refcounted. This means that multiple mods can still have overlapping effects on an actor (using different string tags), but the same mod (using the same string tag in all of its systems) can conflict with itself; avoiding such a scenario is the scripter's responsibility. It also means that we can brazenly delete registrations on leveled actors, so there shouldn't be any zombie data.

In addition to these changes, the "force_remove" methods were renamed to overloads of "remove_all_of" to better reflect a similarity in function; remove calls now use STL algorithms; Feature::affectedActors is now a std::set; and the ShouldCancelDetection function, used by our direct patch to the game's detection code, no longer honors Feature locks (on the admittedly unproven assumption that Papyrus script code will not be executing at the same time as the game's core detection code, whether because they run in the same thread or because the game has locking mechanisms of its own), which should help ensure we don't impact performance in scenes with several actors.