Add files via upload
Uploading work on documentation. We'll be using XML files to hold the content and JavaScript to build it all into a single cohesive page. We can't actually distribute that as a finished product, since Chrome insists on having the worst handling for local XHR imaginable, but we can use it to "bake" finished, non-scripted doc files. Probably.