Add files via upload
Papyrus methods can now have a "notes" section. Modified our templating system to allow HTML content in our XML files to also use templating (e.g. if many methods in a class share an argument and I don't want to copy and paste twenty copies of its description); added a storage area to Papyrus classes for templated content. Fixed the "wrap-if-bare" functionality in the template system; it now properly handles the case of "bare" content that nonetheless contains some HTML tags. Changed how classes and methods are loaded, so that the page can cut down on redundant requests and the code is cleaner.