Experimenting with Lua integration
I want to see if I can bolt a Lua script system onto Skyrim, and use it to provide access to high-value engine hooks (so that I don't have to hardcode everything that everyone could possibly want; I can just give people the means to script the hooks and achieve what they want). Lua is valuable because it's relatively lightweight, cleanly designed, fairly easy to develop with, and most importantly of all: we can run it synchronously, unlike Papyrus, where several script calls are "tasks" that halt script execution until a queueing system in the main thread gets around to them.

Currently, we have a single test function that spins up a Lua VM, makes preparations (inject a function for access to our log file, and inject a table with dummy data), runs a single script, and then tears down the VM. This function is called during the DLL load process. And it works!