Redesigned IForm
Redesigned IForm in an attempt to make it suitable for both forms that are always in memory, and forms that can be removed from memory. As part of this change, IForm subclasses now need to register a factory function with their form type, to ensure that they can be constructed properly (since an IForm subclass can contain extra information (changing its size) that can be used to re-locate the wrapped form if the wrapped form is unloaded and later re-loaded, e.g. the form ID of a reference or the worldspace and grid coordinates of an exterior cell).