Failed attempt at game integration
So the current state of the project is an absolute shambles; this commit is mainly just for backup purposes.

So we have a number of problems here, and one of them is that Skyrim doesn't have just one main thread -- or rather, it *does*, but a lot of tasks that you would expect to run on the main thread can run on it, or on any of two "AI linear task threads." This includes tasks that we need to be able to influence with Lua.

The problem is that it's... really difficult... to get Lua to support any sort of multi-threading. I don't need *concurrency* -- I don't need to be able to run multiple pieces of Lua code at the same time -- but I do need thread-safety, and the few bits of documentation available online have been falling apart completely during testing.

This is complicated by the fact that we're approacing literally everything exactly backwards -- from the bottom up. I have, to date, never worked directly with Win32 threading APIs. My code isn't creating the threads that I'm trying to run Lua on. My code has no idea when those threads start and stop running. My code has no idea which threads wait on which other threads, or when threads synchronize in general. I have no idea when or how Lua handles cleanup for spawned threads/states, so I don't know what will happen if a thread stops and I don't manually kill its lua_State.

Moreover, while Lua 5.1 defines macros that can be leveraged for multi-threading support, some of these are missing in Lua 5.3, and all of them appear to be absent from Lua's official documentation. So, right now, my code crashes the instant it tries to create the VM, because Lua calls a macro that (with my setup) accesses a critical-section, but it doesn't call the macro that (with my setup) initializes that critical-section and prepares it for use.

This means two things: the first problem to try and solve is the macro one; and expect the quality of this project's code to drop from "acceptable" to "outright horrible" until I get a handle on things... if I *can* get a handle on this. I picked a hell of a first Lua project, eh?