Critical fixes; event work
Fixed critical bugs that caused Lua to hang or crash when stopping and restarting the VM: IForm and friends used a static bool to initialize (which is a "once only" thing), and now instead check for the existence of their metatables; lua_close doesn't call Lua's unlock macro, so we need to do that manually or Lua deadlocks when we try to restart it.

Attempted to work on the event system some more. It's supposed to take a unique string identifier for the event registration, along with a function, and to allow unregistration; however, this is currently broken, so events never run. With the restart issues out of the way, we can begin to look into this.