Added (untested) interfaces and methods; improved event/hook code (tested)
Events/hooks now share a helper function for iterating over a list of listeners, which will cut down on duplicated code. In addition, if we decide that we want to attach extra data to listeners (e.g. filters of some kind) such that listeners have to be stored as tables, we now only have to update one piece of code.