IEffectShader work; partial library importing
Added more accessors for IEffectShader, and split the accessors up: pretty sure that templated functions are duplicated in their entirety, so I took advantage of pointer-to-members as arguments and moved all "common" code out of templated function bodies.

Wrote a helper function to import only the safe functions of a specified library. Let's see if it works!