Lua form-pointer abandonment is now implemented
Code that wants to call into Lua now creates a "wrapped pointer" object that uses a thread lock. When that object is destroyed, that means that nothing is running inside of Lua anymore, and we can run cleanup tasks.

One of those cleanup tasks is to loop over every wrapper (IForm) created for forms, and check if the wrapper represents a form type that can unload at run-time; if so, we tell the wrapper to abandon its wrapped pointer. The next time Lua tries to work with it (via IForm::unwrap), the wrapper will use its form ID (and optionally other data) to re-obtain the form pointer (via IForm::resolve).

This has been implemented, tested, and currently works.