Removed old Lua thread-safety code
Thread-safety is now handled by wrapped_lua_pointer, so we no longer need to use LUA_USER_H or the lua_lock and lua_unlock macros. One notable benefit to this is that we're no longer locking and unlocking with every API call that does more than read (e.g. every call to lua_pushnumber and friends), as would be the case if we kept using lua_lock and lua_unlock; instead, we now lock and unlock once for the entire block of code in which a wrapped_lua_pointer instance exists.