Lua test scripts
We now longer execute a test script at a hardcoded path when starting the Lua VM (and we certainly don't abort if the test script file is missing).

All test script code has now been moved into a "formal" add-on. I'm also putting the test add-ons for the manifest system up on Github.