Lua interpreter work
Fixed Lua stack management in each of the hooks and in the safeCall utility function.

Added an event sink to the Lua interpreter, for when an actor begins to cast a spell.

Added a few functions to IActor. Added an API for showing notifications in the UI/HUD.

Fixed some compiler errors stemming from changes made to the RE library.